constructor refactor for transaction factories
