+++ src/index.ts
@@ -21,3 +21,4 @@ export * from "./transactionPayload";
 export * from "./transactionWatcher";
 export * from "./transferTransactionsFactory";
 export * from "./utils";
+export * from "./transactionComputer";
