+++ src/smartcontracts/smartContract.local.net.spec.ts
@@ -14,7 +14,7 @@ import { TransactionsFactoryConfig } from "../transactionsFactories/transactions
 import { SmartContractTransactionsFactory } from "../transactionsFactories/smartContractTransactionsFactory";
 import { TokenComputer } from "../tokens";
 import { promises } from "fs";
-import { TransactionComputer } from "../transaction";
+import { TransactionComputer } from "../transactionComputer";
 
 describe("test on local testnet", function () {
     let alice: TestWallet, bob: TestWallet, carol: TestWallet;
