+++ src/transactionsFactories/smartContractTransactionsFactory.ts
@@ -35,10 +35,10 @@ export class SmartContractTransactionsFactory {
     private readonly tokenComputer: TokenComputer;
     private readonly dataArgsBuilder: TokenTransfersDataBuilder;
 
-    constructor({ config, abi, tokenComputer }: { config: Config; abi?: IAbi; tokenComputer: TokenComputer }) {
-        this.config = config;
-        this.abi = abi;
-        this.tokenComputer = tokenComputer;
+    constructor(options: { config: Config; abi?: IAbi; tokenComputer: TokenComputer }) {
+        this.config = options.config;
+        this.abi = options.abi;
+        this.tokenComputer = options.tokenComputer;
         this.dataArgsBuilder = new TokenTransfersDataBuilder();
     }
